pyproject.toml
[tool.poetry]
name = "fastapi-microblog"
version = "0.1.0"
description = ""
authors = ["darialissi"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.112.0"
sqlalchemy = "^2.0.32"
uvicorn = "^0.30.6"
httpx = "^0.27.0"
asyncpg = "^0.29.0"
fastapi-cache2 = {extras = ["redis"], version = "^0.2.2"}
pydantic-settings = "^2.4.0"
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
pyjwt = {extras = ["crypto"], version = "^2.9.0"}
python-multipart = "^0.0.9"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.2"
black = "^24.8.0"
isort = "^5.13.2"
flake8-pyproject = "^1.2.3"

[tool.black]
line-length = 120
target_version = ["py311"]

[tool.isort]
profile = "black"
py_version = 311

[tool.flake8]
max-line-length = 120

[tool.pytest.ini_options]
pythonpath = [".", "src"]
asyncio_mode = "auto"
filterwarnings = "ignore::DeprecationWarning"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
